requirements.txt
# NOTE: Install jax manually from online instructions, after installing these requirements
# jax>=0.2.26
# jaxlib>=0.1.74
# cuda 12:
# pip install --upgrade "jax[cuda12_pip]" -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
# cuda 11:
# pip install --upgrade "jax[cuda11_pip]" -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
distrax
evosax
flask
flax
gradio
gymnax
hydra-core
imageio
jinja2
numpy
optax
pandas
safetensors
submitit
tabulate==0.9.0
tensorboard
tensorboardX
tensorflow-probability[tf]
tqdm==4.66.2
wandb
